covid/requirements.txt
streamlit
pandas
plotly
snowflake-connector-python
python-dotenv
